backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
fastembed
numpy
sqlalchemy
pgvector
python-dotenv
pydantic